refactor(api): split routes/workflows.rs into per-concern modules
routes/workflows.rs (3.9k lines) bundled workflow, trigger, schedule, cron,
and template handlers in one file. Convert it into a routes/workflows/
module directory grouped by domain. `workflows::router` (the only
externally-referenced symbol) is preserved via re-export.

- mod.rs keeps router(), the import block, every non-handler helper
  (workflow_to_json, parse_step_mode/parse_error_mode/parse_step_session_mode,
  parse_input_schema, operator_pause_row_json, trigger_to_json,
  normalize_pattern_json, parse_cron_job_id, cron_job_to_schedule_json,
  the cron_* response helpers, workflow_run_input_string), the `mod`/`pub use`
  web, and the test module.
- Handler submodules: workflow (CRUD + runs + save-as-template), triggers,
  schedules, cron, templates. Each handler/DTO is moved verbatim; submodules
  reach shared helpers via `use super::*`.

Pure structural change: every moved item is byte-identical (verified by line
conservation — 0 missing, 0 extra beyond re-export scaffolding); no handler,
signature, or behavior changed. Keeping all helpers in mod.rs (vs scattering
them) means no cross-module private reference can break — verified by static
analysis (0 parent->child, 0 sibling->sibling private refs).

Verification: line-conservation + brace/paren balance + cross-boundary
reference scan + detect-secrets re-scan (unchanged). No native cargo/rustfmt
in this environment — relying on CI.
